v1.2.0 Feature - Parameterizable Messages
This feature includes:
- Display numbers, currency, dates and times for different locales.
- Pluralize labels in strings.
- Support variables in message.
- Message format is strictly implemented by ICU standards.
- Locale data in nested JSON format are supported.

New React Hook:
- useTranslation

This hook receives an object with several dictionaries as parameter,
each dictionary must contain the definition (according to locale) for
each of the terms used by the application that uses this lib.
A second parameter can be optionally passed to customize number and
date formatting.

This hook returns a function capable of translating a Term according
to the current set locale. The function takes two parameters, the first
is the term to be translated, the second (optional) is a key/value object
used in case of parameterizable messages.

This feature makes use of 'intl-messageformat' lib. For more details see:
https://github.com/formatjs/intl-messageformat
This lib is based on the specification ECMAScript Internationalization API 4.0
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Intl